Reshuffle so loading frame is in right place
